Added BluetoothContext for alternative use cases. Updated README accordingly. @TODO: add more tests
