+++ src/BluetoothProvider.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+import { BluetoothNotificationsHookOptions } from "./types/Bluetooth";
+import { BluetoothContext } from "./BluetoothContext";
+import { useBluetoothNotifications } from "./useBluetoothNotifications";
+
+export const BluetoothProvider: React.FC<React.PropsWithChildren<BluetoothNotificationsHookOptions>> = ({
+  children,
+  ...props
+}) => {
+  const bluetoothState = useBluetoothNotifications(props);
+  return <BluetoothContext.Provider value={bluetoothState}>{children}</BluetoothContext.Provider>;
+};
+
+export default BluetoothProvider;
