+++ src/useBluetooth.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import { BluetoothConnections } from "./types/Bluetooth";
+import BluetoothContext from "./BluetoothContext";
+
+export function useBluetooth() {
+  const context = React.useContext<BluetoothConnections>(BluetoothContext);
+  if (context === undefined) {
+    throw new Error("useBluetooth must be used within a BluetoothProvider");
+  }
+  return context;
+}
+
+export default useBluetooth;
